Substitute preliminary coGN submission with new (refactored, de-anonymized) one
